Make horizontal borders the size of the cell in themes.
Because that way it makes more sense.
Also just hardcode 2 as the offset for score counter, because using
the width of the cell for that didn't actually make sense.
Because it starts with `border.top_left`, which is 1 char wide and only
then there are `border.horizontal`s which are 2 chars wide.